Fixed Ranks & Claim Deletion
I fixed a small qwerk with the claim deletion where it was running too soon. I also re-ordered the ranks. Sab and I looked over the code for instances of the use of the EnumRank class that needed to be updated due to the re-ordering. We believe that we checked everywhere, however, another set of eyes wouldn't hurt.